+++ app/src/main/java/com/bee/user/rest/ApiService.java
@@ -206,7 +206,7 @@ public interface ApiService {
      * 获取购物车信息
      */
     @GET(HttpRequest.getCart)
-    Observable<BaseResult<List<ChartBean>>> getCart( @Query("receiveAddressId") Integer receiveAddressId, @Query("storeIds") List<String> storeIds);
+    Observable<BaseResult<List<ChartBean>>> getCart( @Query("receiveAddressId") Long receiveAddressId, @Query("storeIds") List<String> storeIds);
     /**
      * 获取购物车某个购物项
      */
